Fix bugs multi-amplicon and small OTU tables
With multiple primer combinations analyzed in
one project, there will always be some unexpected
combinations with very few reads, which ultimately
can result in empty or very small OTU tables.
There were still some problems with this case.
The R import script also needed adjustments to work
with very small OTU tables containing only single samples
and/or single OTUs.

+++ workflow/scripts/usearch/make_otutab.sh
@@ -37,11 +37,15 @@ if [[ "$program" == "vsearch" ]]; then
         "$@"
 
     # SAM -> BAM
+    if [ -s "$otus" ]; then
     rm -f "$otus".fai "$bam".bai
-    samtools view -T "$otus" -b "$sam" |
-        samtools sort -@ $threads > "$bam"
-    rm $sam "$otus".fai
-    samtools index "$bam"
+        samtools view -T "$otus" -b "$sam" |
+            samtools sort -@ $threads > "$bam"
+        rm $sam "$otus".fai
+        samtools index "$bam"
+    else
+        echo -n > "$bam"
+    fi
 
 elif [[ "$program" == "usearch" ]]; then
     uniques=$(mktemp ${uniques_compr%.fasta.zst}.XXXXXX.fasta)
